pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "textbaker"
version = "0.1.0"
description = "Synthetic Text Dataset Generator for OCR Training"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Ramkrishna Acharya"}
]
maintainers = [
    {name = "Ramkrishna Acharya"}
]
keywords = ["ocr", "dataset", "synthetic", "text", "image-generation", "machine-learning"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "PySide6>=6.4.0",
    "loguru>=0.6.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-qt>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=0.990",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/q-viper/text-baker"
Documentation = "https://q-viper.github.io/text-baker"
Repository = "https://github.com/q-viper/text-baker"
Issues = "https://github.com/q-viper/text-baker/issues"

[project.scripts]
textbaker = "textbaker.cli:run"

[project.gui-scripts]
textbaker-gui = "textbaker.cli:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["textbaker*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument
    "ARG002", # unused method argument
]

[tool.ruff.lint.isort]
known-first-party = ["textbaker"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["textbaker"]
branch = true
omit = [
    "textbaker/__main__.py",
    "textbaker/app/*",  # GUI tests require display
    "textbaker/widgets/*",  # GUI tests require display
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
